# apis/openapi.yaml
openapi: 3.0.3
info:
  title: Inventory & Storage API
  version: "1.0.0"
  description: >
    The WMS-tier authoritative record of what is held where, and what
    portion of it is usable, for the warehouse-systems platform. This
    bounded context implements Amazon-style chaotic (random) stow: there is
    no fixed product location — an inbound item may be stowed into any free
    bin, and this service records the exact bin it landed in. It supplies
    "stock reality" to the Work Planning bounded context and makes
    allocation a revocable reservation, so a failed physical pick or
    delivery never strands an order: a reservation can always be revoked
    and re-satisfied from a different physical holding.


    Its aggregates are **StockUnit** (a quantity of a SKU at a specific
    bin — every physical item has exactly one known bin, or is flagged
    Unlocated), **Bin/Location** (a coded slot with fixed capacity; any SKU
    may occupy any free bin, chaotic-storage style), and **Reservation** (a
    revocable binding of quantity to demand, with a timeout). "Usable"
    inventory — on-hand minus active reservations minus held/unlocated
    stock — is what constrains what this service will let you reserve, not
    raw on-hand quantity.


    Out of this service's boundary: it does not pick, pack, ship, route
    associates, or manage stations/tasks/labor (that is Work Planning /
    fulfillment-execution); and it does not decide *what* to receive or
...
    Problem Details.
...
servers:
...
paths:
...
  /stock/receive:
    post:
      operationId: receiveStock
      summary: Acknowledge inbound stock for a SKU
      description: >
        Stages a quantity of a SKU as received into the building
        (publishes StockReceived) without assigning it to a bin yet. This
        does NOT create a StockUnit aggregate: a StockUnit requires both an
        item-scan and a location-scan by construction (the
        stow-requires-both invariant), and a staged receipt has no bin yet.
        The durable, addressable record only starts at StowStock — that is
        why this operation returns 202 Accepted rather than 201 Created:
        there is no resource here to hand back a Location for.
      tags: [Stock]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveStockRequest"
            example:
              sku: SKU-1
              quantity: 10
      responses:
        "202":
          description: >
            Receipt accepted and staged; not yet stowed to a bin.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StagedReceipt"
              example:
                sku: SKU-1
                quantity: 10
                receivedAt: "2026-08-22T14:20:00Z"
        "400":
          description: >
            Malformed JSON body, SKU missing/empty, or the Idempotency-Key
            header is absent (ADR 0018).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                emptySku:
                  summary: SKU missing or empty
                  value:
                    type: https://errors.inventory-storage.warehouse-systems.dev/empty-sku
                    title: SKU must not be empty
                    status: 400
                    detail: sku must not be empty
                    instance: /stock/receive
                idempotencyKeyRequired:
                  summary: Idempotency-Key header missing (ADR 0018)
                  value:
                    type: https://errors.inventory-storage.warehouse-systems.dev/idempotency-key-required
                    title: Idempotency-Key header is required
                    status: 400
                    detail: This endpoint creates a new resource on every call and requires a caller-supplied Idempotency-Key header so a retried request is never applied twice.
                    instance: /stock/receive
        "422":
          description: >
            Quantity present but not a positive number, or the
            Idempotency-Key was already used with a different request body
            (ADR 0018).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                zeroQuantity:
                  summary: Quantity not positive
                  value:
                    type: https://errors.inventory-storage.warehouse-systems.dev/zero-quantity
                    title: Quantity must be greater than zero
                    status: 422
                    detail: quantity must be greater than zero
                    instance: /stock/receive
                idempotencyKeyReused:
                  summary: Idempotency-Key reused with a different request body (ADR 0018)
                  value:
                    type: https://errors.inventory-storage.warehouse-systems.dev/idempotency-key-reused
                    title: Idempotency-Key was already used with a different request
                    status: 422
                    detail: The Idempotency-Key header on this request was already used with a request that had a different body. Use a new Idempotency-Key for a genuinely different request.
                    instance: /stock/receive
        "500":
          $ref: "#/components/responses/InternalError"
...
  /reservations/{id}:
    delete:
      operationId: revokeReservation
      summary: Revoke a reservation
      description: >
        Cancels an ACTIVE reservation and returns its full reserved
        quantity to usable inventory on every StockUnit it had allocated
        from. This is the mechanism that makes reservations safe against
        physical delivery failure: revoke, then reserve again — the second
        reservation is free to draw from a different physical holding.
        Idempotent in effect (a resource that no longer exists or is
        already resolved simply can't be revoked again), though a second
        call on an already-resolved reservation is reported as 409, not a
        silent 204.
      tags: [Reservations]
      parameters:
        - $ref: "#/components/parameters/ReservationId"
      responses:
        "204":
          description: Reservation revoked; quantity returned to usable.
        "404":
          description: No reservation exists with this id.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.inventory-storage.warehouse-systems.dev/reservation-not-found
                title: Reservation not found
                status: 404
                detail: reservation not found
                instance: /reservations/does-not-exist
        "409":
          description: >
            The reservation is already resolved (confirmed, revoked, or
            expired), or a racing request modified the same reservation
            first (ADR 0019 — re-fetch and retry).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              examples:
                reservationAlreadyResolved:
                  summary: Reservation is already resolved
                  value:
                    type: https://errors.inventory-storage.warehouse-systems.dev/reservation-already-resolved
                    title: Reservation is already resolved
                    status: 409
                    detail: reservation is already resolved (confirmed, revoked, or expired)
                    instance: /reservations/res-a1b2c3d4-0000-0000-0000-000000000001
                concurrentModification:
                  summary: A racing request modified the same reservation first (ADR 0019)
                  value:
                    type: https://errors.inventory-storage.warehouse-systems.dev/concurrent-modification
                    title: The resource was modified by another request; re-fetch the latest version and retry
                    status: 409
                    detail: the resource was modified by another request; re-fetch the latest version and retry
                    instance: /reservations/res-a1b2c3d4-0000-0000-0000-000000000001
        "500":
          $ref: "#/components/responses/InternalError"
...
  /inventory/{sku}/usable:
    get:
      operationId: getUsableInventory
      summary: Get usable inventory for a SKU
      description: >
        Returns the usable-inventory read model for a SKU: on-hand minus
        active reservations minus held/unlocated stock, summed across
        every StockUnit for that SKU. This is a projection, not a stored
        aggregate — it is what constrains ReserveStock, since total
        on-hand is not what's available to promise to new demand.
        Side-effect-free.
      tags: [Inventory]
      parameters:
        - $ref: "#/components/parameters/Sku"
      responses:
        "200":
          description: Usable quantity for the SKU (0 if the SKU is unknown).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsableInventory"
              example:
                sku: SKU-1
                usable: 7
        "400":
          description: SKU path segment missing/empty.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.inventory-storage.warehouse-systems.dev/empty-sku
                title: SKU must not be empty
                status: 400
                detail: sku must not be empty
                instance: /inventory//usable
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    ReservationId:
      name: id
      in: path
      required: true
      description: The reservation's id, formatted `res-<uuid>`.
      schema:
        type: string
        example: res-a1b2c3d4-0000-0000-0000-000000000001
    Sku:
      name: sku
      in: path
      required: true
      description: The stock-keeping unit identifier.
      schema:
        type: string
        example: SKU-1
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: >
        Caller-generated unique key (a UUID per logical submit) that makes
        a retried request safe (ADR 0018). Required whenever the service
        runs with Postgres — i.e. in every deployment — and answered with
        400 `idempotency-key-required` when absent; a replay with the same
        key and the same body returns the original response, while the same
        key with a different body is 422 `idempotency-key-reused`. The
        in-memory development mode (no DATABASE_URL) does not enforce it,
        which is why the header is not marked `required` here (the
        Schemathesis contract run boots that mode).
      schema:
        type: string
        example: 5b0f6a0e-6a41-4c63-9f33-0c8f5c1d2a77
...
  responses:
    InternalError:
      description: An unexpected internal error occurred.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          example:
            type: https://errors.inventory-storage.warehouse-systems.dev/internal-error
            title: An unexpected internal error occurred
            status: 500
            detail: unexpected adapter failure
            instance: /stock/stow
  schemas:
    Problem:
      type: object
      description: >
        RFC 7807 (Problem Details for HTTP APIs) error shape, used for
        every non-2xx response across this service. `type` is
        `https://errors.inventory-storage.warehouse-systems.dev/<slug>`;
        every slug is documented as a response example on the operation(s)
        that can return it, and the mapping from typed domain error to
        (status, slug, title) lives in one lookup table
        (`problemCatalog` in internal/adapters/inbound/http/errors.go,
        ADR 0005).
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          description: >
            A URI identifying the error category (does not need to
            resolve). Unique per distinct error category in this service.
          example: https://errors.inventory-storage.warehouse-systems.dev/reservation-not-found
        title:
          type: string
          description: Fixed, human-readable summary of the error category.
          example: Reservation not found
        status:
          type: integer
          format: int32
          description: Duplicates the HTTP status code as an integer (RFC 7807 requirement).
          example: 404
        detail:
          type: string
          description: The specific error message for this occurrence.
          example: reservation not found
        instance:
          type: string
          description: The request path that produced this error.
          example: /reservations/does-not-exist
    ReceiveStockRequest:
      type: object
      required: [sku, quantity]
      properties:
        sku:
          type: string
          minLength: 1
          description: Stock-keeping unit identifier for the inbound goods.
          example: SKU-1
        quantity:
          type: integer
          format: int64
          minimum: 1
          description: Quantity received. Must be a positive integer.
          example: 10
    StagedReceipt:
      type: object
      properties:
        sku:
          type: string
          example: SKU-1
        quantity:
          type: integer
          format: int32
          example: 10
        receivedAt:
          type: string
          format: date-time
          description: Timestamp the receipt was staged, RFC 3339.
          example: "2026-08-22T14:20:00Z"
...
    UsableInventory:
      type: object
      properties:
        sku:
          type: string
          example: SKU-1
        usable:
          type: integer
          format: int32
          description: On-hand minus active reservations minus held/unlocated stock, summed across all StockUnits for this SKU.
          example: 7
...